requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
gitpython==3.1.42 